Implement multi-state for rhomb drawer
This application will be used to emulate cellular automata on
the Penrose tiling. In CA, cells (which correspond to PT tiles)
are each in some state.

This patch adds ability to load and display images for different
states. States in the universe will be implemented later.

Also, some refactoring of rhomb loader has been made.