Implement dumping rhomb icons to sources
It is often desirable to adjust the existing tile icons
for a new icon set. It may require using extended features,
like icons depending on von Neumann neighbourhood.

This patch adds the ability to dump currently used icon set
to a compatible source file. It supports dumping into simple
source (S vertex) or von Neumann extension source (S5 vertex).

The type of the target source doesn't have to be the same as
the original source from which the icon set was loaded.

Note, however, that some details are lost when dumping a von
Neumann icon set into a simple source.

Currently it dumps into duplex sources regardless of the
original sources' type.